Ai_agent/requirements.txt
Flask==2.2.3
requests==2.28.2
retell-sdk==0.1.0
